fix: preserve knowledge review page functionality
